factory/raspberry/controller/requirements.txt
rpi_ws281x
adafruit-circuitpython-neopixel
colorutils
RPi.GPIO